move tensor storage to a separate file, use delegation instead of inheritance
