+++ oneflow/api/python/dlpack/converter.cpp
@@ -18,6 +18,7 @@ limitations under the License.
 #include "oneflow/api/python/of_api_registry.h"
 #include "oneflow/core/common/data_type.h"
 #include "oneflow/core/eager/eager_blob_object.h"
+#include "oneflow/core/eager/tensor_storage.h"
 #include "oneflow/core/framework/tensor.h"
 #include "oneflow/core/framework/device.h"
 #include "oneflow/core/framework/tensor_util.h"
@@ -105,7 +106,7 @@ Maybe<one::Tensor> fromDLPack(const DLManagedTensor* src) {
   };
 
   size_t array_size_in_bytes = shape.elem_cnt() * GetSizeOfDataType(dtype);
-  auto tensor_data = std::make_shared<vm::OutsideVmTensorStorage>();
+  auto tensor_data = std::make_shared<vm::TensorStorage>(false);
   tensor_data->set_blob_dptr(
       std::unique_ptr<char, std::function<void(char*)>>(static_cast<char*>(dl_tensor.data), Free),
       array_size_in_bytes);
