+++ oneflow/core/eager/eager_blob_object.cpp
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 #include "oneflow/core/eager/eager_blob_object.h"
+#include "oneflow/core/eager/tensor_storage.h"
 #include "oneflow/core/vm/allocator.h"
 #include "oneflow/core/framework/to_string.h"
 #include "oneflow/core/framework/shut_down_util.h"
@@ -109,6 +110,37 @@ Maybe<bool> EagerBlobObject::TryAllocateBlobBodyMemory(vm::Allocator* allocator)
     return true;
   }
   return false;
+}
+
+const void* EagerBlobObject::raw_dptr() const {
+  char* ptr = tensor_storage_->blob_dptr();
+  if (tensor_storage_->blob_bytes() > 0) { CHECK_NOTNULL(ptr); }
+  return ptr + storage_offset_ * GetSizeOfDataType(data_type_);
+}
+
+Maybe<void> EagerBlobObject::DeallocateBlobDataPtr() {
+  tensor_storage_->Release();
+  return Maybe<void>::Ok();
+}
+
+void EagerBlobObject::RegisterStorageDeleteHook(const std::function<void()>& hook) {
+  tensor_storage_->RegisterStorageDeleteHook(hook);
+}
+
+const Optional<Symbol<::oneflow::Stream>>& EagerBlobObject::producer_stream() const {
+  return tensor_storage_->producer_stream();
+}
+
+Maybe<void> EagerBlobObject::init_producer_stream(Symbol<::oneflow::Stream> producer_stream) {
+  return tensor_storage_->init_producer_stream(producer_stream);
+}
+
+const Optional<Symbol<::oneflow::Stream>>& EagerBlobObject::last_used_stream() const {
+  return tensor_storage_->last_used_stream();
+}
+
+void EagerBlobObject::set_last_used_stream(Symbol<::oneflow::Stream> last_used_stream) {
+  tensor_storage_->set_last_used_stream(last_used_stream);
 }
 
 }  // namespace vm
