+++ oneflow/core/eager/eager_blob_object.h
@@ -40,77 +40,6 @@ class MutLocalTensorMeta;
 
 namespace vm {
 
-class TensorStorage {
- public:
-  TensorStorage()
-      : blob_bytes_(0),
-        non_pod_allocator_(std::make_unique<MemoryAllocator>()),
-        producer_stream_(NullOpt),
-        last_used_stream_(NullOpt) {}
-
-  virtual ~TensorStorage() {
-    for (const auto& hook : storage_delete_hooks_) { hook(); }
-  }
-
-  virtual bool is_allocated_in_vm() const = 0;
-
-  size_t blob_bytes() const { return blob_bytes_; }
-
-  char* blob_dptr() { return blob_dptr_.get(); }
-
-  MemoryAllocator* non_pod_allocator() { return non_pod_allocator_.get(); }
-
-  void set_blob_dptr(std::unique_ptr<char, std::function<void(char*)>>&& blob_dptr, size_t bytes) {
-    blob_dptr_ = std::move(blob_dptr);
-    blob_bytes_ = bytes;
-  }
-
-  const Optional<Symbol<::oneflow::Stream>>& producer_stream() const { return producer_stream_; }
-  Maybe<void> init_producer_stream(Symbol<::oneflow::Stream> producer_stream) {
-    CHECK_OR_RETURN(!producer_stream_.has_value());
-    producer_stream_ = producer_stream;
-    return Maybe<void>::Ok();
-  }
-
-  const Optional<Symbol<::oneflow::Stream>>& last_used_stream() const { return last_used_stream_; }
-  void set_last_used_stream(Symbol<::oneflow::Stream> last_used_stream) {
-    last_used_stream_ = last_used_stream;
-  }
-
-  void Release() {
-    non_pod_allocator_.reset();
-    blob_dptr_.reset();
-  }
-
-  void RegisterStorageDeleteHook(const std::function<void()>& hook) {
-    storage_delete_hooks_.emplace_back(hook);
-  }
-
- private:
-  size_t blob_bytes_;
-  std::unique_ptr<char, std::function<void(char*)>> blob_dptr_;
-  std::unique_ptr<MemoryAllocator> non_pod_allocator_;
-  Optional<Symbol<::oneflow::Stream>> producer_stream_;
-  Optional<Symbol<::oneflow::Stream>> last_used_stream_;
-  std::vector<std::function<void()>> storage_delete_hooks_;
-};
-
-class InsideVmTensorStorage : public TensorStorage {
- public:
-  InsideVmTensorStorage() = default;
-  ~InsideVmTensorStorage() = default;
-
-  bool is_allocated_in_vm() const override { return true; }
-};
-
-class OutsideVmTensorStorage : public TensorStorage {
- public:
-  OutsideVmTensorStorage() = default;
-  ~OutsideVmTensorStorage() = default;
-
-  bool is_allocated_in_vm() const override { return false; }
-};
-
 class EagerBlobObject final : public user_op::Tensor,
                               public user_op::TensorDesc,
                               public std::enable_shared_from_this<EagerBlobObject> {
@@ -152,25 +81,15 @@ class EagerBlobObject final : public user_op::Tensor,
   ShapeView shape_view() const override { return shape(); }
   MutShapeView mut_shape_view() override;
   const MemoryCase& mem_case() const override { return *mem_case_; }
-  const void* raw_dptr() const override {
-    char* ptr = tensor_storage_->blob_dptr();
-    if (tensor_storage_->blob_bytes() > 0) { CHECK_NOTNULL(ptr); }
-    return ptr + storage_offset_ * GetSizeOfDataType(data_type_);
-  }
+  const void* raw_dptr() const override;
   void* mut_raw_dptr() override { return const_cast<void*>(raw_dptr()); }
 
   void set_storage_offset(const int64_t offset);
 
   // Returns true if allocate successfully.
   Maybe<bool> TryAllocateBlobBodyMemory(vm::Allocator* allocator);
-  Maybe<void> DeallocateBlobDataPtr() {
-    tensor_storage_->Release();
-    tensor_storage_.reset(new InsideVmTensorStorage());
-    return Maybe<void>::Ok();
-  }
-  void RegisterStorageDeleteHook(const std::function<void()>& hook) {
-    tensor_storage_->RegisterStorageDeleteHook(hook);
-  }
+  Maybe<void> DeallocateBlobDataPtr();
+  void RegisterStorageDeleteHook(const std::function<void()>& hook);
 
   Maybe<LocalDepObject*> compute_local_dep_object() const {
     CHECK_NOTNULL_OR_RETURN(compute_local_dep_object_.get());
@@ -179,19 +98,11 @@ class EagerBlobObject final : public user_op::Tensor,
 
   std::shared_ptr<TensorStorage>& tensor_storage() { return tensor_storage_; }
 
-  const Optional<Symbol<::oneflow::Stream>>& producer_stream() const {
-    return tensor_storage_->producer_stream();
-  }
-  Maybe<void> init_producer_stream(Symbol<::oneflow::Stream> producer_stream) {
-    return tensor_storage_->init_producer_stream(producer_stream);
-  }
+  const Optional<Symbol<::oneflow::Stream>>& producer_stream() const;
+  Maybe<void> init_producer_stream(Symbol<::oneflow::Stream> producer_stream);
 
-  const Optional<Symbol<::oneflow::Stream>>& last_used_stream() const {
-    return tensor_storage_->last_used_stream();
-  }
-  void set_last_used_stream(Symbol<::oneflow::Stream> last_used_stream) {
-    tensor_storage_->set_last_used_stream(last_used_stream);
-  }
+  const Optional<Symbol<::oneflow::Stream>>& last_used_stream() const;
+  void set_last_used_stream(Symbol<::oneflow::Stream> last_used_stream);
 
   std::shared_ptr<const Shape> shape_ptr() const;
   std::shared_ptr<const Stride> stride_ptr() const;
